Add direct_proxy configuration variable
When set to the name of an upstream, the direct_proxy variable causes
ALL requests to be proxied (so there is no try_file directive).

This is useful, for example, when proxying to a web server which handles
all static content or when the nginx reverse-proxy doesn't have access
to the origin server's root directory.

Example:

direct_proxy: apache
upstreams:
     - name: apache
       servers: ['apache_upstream_server']